adds delay for starting the protocol

+++ pkg/config/config.go
@@ -9,7 +9,7 @@ type Configuration struct {
 	CreateDelay float32
 
 	// a number in (0,1) describing how aggressive is the create_delay adjusting mechanism, large = aggressive
-	StepSize float32
+	StepSize float64
 
 	// delay after initializing a sync with other processes
 	SyncInitDelay float32
